- Refactored some GM code, added multiple unit tests for actor searching, - Merge with delta3d CVS - Adding Unit Tests for the Log Module - fixed an infinite recursion and some errors in actor searching - renamed unit test files to be lowercase - fixed invokable tests and added a RemoveAllDrawables method to Scene - many Game manager updates and test additions, mainly related to processing messages - extended process message tests - Added ActorParameter types, added a static method to create parameters based on type, fixed some process message things, changes GetActorUpdate to PopulateActorUpdate, updated unit tests for message passing. - added publish actor tests - Modified to support the GetRealActor on a actoractorproperty. Wrote unit tests to ensure it worked. - added message cloning and some tests for the rules component - Wrote unit tests for the actor msg parameter. - Added Unit tests for enumeration class - Fixed formating in logtests - Added Unit Tests for the exception class - Added comments and formating to log and enumeration tests - Added partially working message listener code and timers - Game manager merged and working in linux - Game Manager sets the GameManager properties on GameActorProxies in add and remove - Removed references to the GameManager dir in DELTA_ROOT. - Changing < enum test.  The previous test should have failed because they were equal - component add/remove tests added - Added const version of GetParameter to base message class. -Tests for stringutils (Found some bugs). The following will be sent in email, but is here also for history's sake TestTrim() 	-The tests for an empty string fails.  This was commented out because the call to trim will crash the program 	-Test for char<space> fails.  Not taking off ending space 	-Test for char<tab> fails.  Not taking off the ending tab 	-Test for <space><tab><space>12345<space><tab><space> fails. Strips off first space and nothing else.
TestToString:
	-Test for converting a double or float with value 0.0 returns 0.  Should it return "0.0"?
- Added more message registration tests, fixed trim in string utils, fixed a warning in stringutilstests.cpp, debugged message registration tests
- Made message unique id's default to empty, added tests for message default values, fixed an error with proxy tests
- pulling out comments relating to trim() failures because they are fixed
- unit tests for the deprecation manager.  Not much here, just made some some functions, deprecated them, called them and made sure the warnings get printed
- fixed comment in stringutils.cpp
- added more tests for message registration.  They currently fail.
- Removed detail scale map from rendering.  This made the terrain look really weird with little spots all over the terrain.  Now that there are good texture colorations, it was not needed.
- Added a new height-based color texture decorator.  Fixes several bugs in the LCC analyzer.  Refactored the vegetation placement and calculations.  There are still issues with the vegetation placement, but there have been several improvements.
- Removed some console printing messages for Terrain.
- Fixed max vegetation cap.
- Pre-merge for v1.1 with terrain and Game Manager(Alpha)
- fixed engine property types compile problem
- Merged BMH testterrain with NPS cvs.
- Fixed a duplicate function call in testActorProxy.
- Fixed message registration, delete bug in gamemanager, disable unregistraition of preframe to temporarily get around a crash, and fixed errors in the unit tests.
- fixed a unit test problem and made the proxy bezier tests not crash
- Fixed line endings for merge

+++ trunk/delta3d/utilities/editor/src/dynamicactorcontrol.cpp
@@ -103,7 +103,7 @@ namespace dtEditQt
             
             std::vector<osg::ref_ptr<dtDAL::ActorProxy> > proxies = GetActorProxies(myProperty->GetDesiredActorClass());
 
-            myProperty->SetValue(index < proxies.size() ? proxies[index].get() : NULL);
+            myProperty->SetValue((unsigned int)index < proxies.size() ? proxies[index].get() : NULL);
             dataChanged = true;
          }
       }
